+++ lib/rufo.rb
@@ -3,7 +3,14 @@
 module Rufo
   class Bug < StandardError; end
 
-  class SyntaxError < StandardError; end
+  class SyntaxError < StandardError
+    attr_reader :lineno
+
+    def initialize(message, lineno)
+      super(message)
+      @lineno = lineno
+    end
+  end
 
   def self.format(code, **options)
     Formatter.format(code, **options)
@@ -14,6 +21,7 @@ require_relative "rufo/command"
 require_relative "rufo/logger"
 require_relative "rufo/dot_file"
 require_relative "rufo/settings"
+require_relative "rufo/parser"
 require_relative "rufo/formatter"
 require_relative "rufo/version"
 require_relative "rufo/file_finder"