+++ lib/rufo/command.rb
@@ -51,8 +51,8 @@ class Rufo::Command
     print(result) if !@want_check
 
     code == result ? CODE_OK : CODE_CHANGE
-  rescue Rufo::SyntaxError
-    logger.error("Error: the given text is not a valid ruby program (it has syntax errors)")
+  rescue Rufo::SyntaxError => e
+    logger.error("STDIN is invalid code. Error on line:#{e.lineno} #{e.message}")
     CODE_ERROR
   rescue => ex
     logger.error("You've found a bug!")
@@ -96,10 +96,10 @@ class Rufo::Command
 
     begin
       result = format(code, @filename_for_dot_rufo || File.dirname(filename))
-    rescue Rufo::SyntaxError
+    rescue Rufo::SyntaxError => e
       # We ignore syntax errors as these might be template files
       # with .rb extension
-      logger.warn("Error: #{filename} has syntax errors")
+      logger.warn("#{filename}:#{e.lineno} #{e.message}")
       return CODE_ERROR
     end
 
@@ -113,8 +113,8 @@ class Rufo::Command
 
       return CODE_CHANGE
     end
-  rescue Rufo::SyntaxError
-    logger.error("Error: the given text in #{filename} is not a valid ruby program (it has syntax errors)")
+  rescue Rufo::SyntaxError => e
+    logger.error("#{filename}:#{e.lineno} #{e.message}")
     CODE_ERROR
   rescue => ex
     logger.error("You've found a bug!")